Added backend logic for taskPoint accumulation

+++ server/handlers/questHandlers.js
@@ -17,15 +17,34 @@ const createQuest = async(req,res) => {
     try{
         const quest = {ownerId: req.params.ownerId, _id: uuidv4(), ...req.body, createdAt: date};
         await client.connect();
+
+        // 100 points of karma per difficulty level to be paid by owner
+        const karma = quest.difficulty * 100;
         
         const db = await client.db("BootCamp_Final_Project");
         console.log("database connected!");
 
-        const questInserted = await db.collection("quests").insertOne(quest);
+        const user = await db.collection("users").findOne({_id: quest.ownerId});
+        
+        // If owner has insufficient karma points to create a quest with this difficulty level - reject the request.
+        if(user.karma < karma){        
+            return res.status(400).json({status:400, data:quest, message: "ERROR: Insufficient karma points to create this quest."});   
+        }
+
+        // If owner has sufficient karma points - deduct points from user.
+        const ownerUpdated = await db.collection("users").updateOne({_id: quest.ownerId}, {$inc: {karma: -karma}, $set: {updatedAt: date}});
+
+        if(ownerUpdated  === null){
+            return res.status(400).json({status:400, message: "ERROR: Owner could not be updated."}); 
+        }
+
+        // Create new quest
+        const questInserted = await db.collection("quests").insertOne({...quest, karma});
 
         questInserted
         ? res.status(201).json({status:201, data:quest, message: "SUCCESS: New Quest created."})
-        : res.status(500).json({status:500, data:quest, message: "ERROR: Internal server error."});   
+        : res.status(400).json({status:400, message: "ERROR: Unable to create quest."}); 
+
     }catch(err){
         console.log(err);
         res.status(500).json({status:500, data:null, message: `ERROR: Internal server error.`});
@@ -46,7 +65,15 @@ const addQuestParticipant = async(req,res) => {
         const db = await client.db("BootCamp_Final_Project");
         console.log("database connected!");
 
-        // TO DO: Reduce participant slots on quest
+        // Backend validation to verify if slot are available 
+        const quest = await db.collection("quests").findOne({_id: req.params.id});
+
+        // Throw error if no slots are available
+        if(quest.participants === 0 ){
+            return res.status(400).json({status:400, data:null, message: "ERROR: Quest is full. Unable to add user to quest."});   
+        }
+
+        // Add user id to the quest
         const questUpdated = await db.collection("quests").updateOne({_id: req.params.id}, { $set: {participants: participants,  updatedAt: date}, $push: {participantIds:  participant }});
 
         questUpdated
@@ -70,11 +97,22 @@ const deleteQuest = async(req,res) => {
         const db = await client.db("BootCamp_Final_Project");
         console.log("database connected!");
 
+        // Return owners karma points as quest was not completed.
+        const quest = await db.collection("quests").findOne({_id: req.params.id});
+        const ownerUpdated = await db.collection("users").updateOne({_id: quest.ownerId}, {$inc: {karma: quest.karma}, $set: {updatedAt: date}});
+
+        // If owner was not updated, throw error
+        if(ownerUpdated  === null){
+            return res.status(400).json({status:400, data:quest, message: "ERROR: Owner's karma points were not returned!"}); 
+        }
+
+        // Delete quest
         const questDeleted = await db.collection("quests").deleteOne({_id: req.params.id});
 
         questDeleted
-        ? res.status(201).json({status:200, message: "SUCCESS: Quest deleted succesfully."})
-        : res.status(500).json({status:500,  message: "ERROR: Internal server error."});   
+        ? res.status(200).json({status:200, data:quest, message: "SUCCESS: Quest deleted successfully."})
+        : res.status(400).json({status:400, message: "ERROR: Unable to delete quest."})
+
     }catch(err){
         console.log(err);
         res.status(500).json({status:500, message: `ERROR: Internal server error.`});
@@ -94,19 +132,52 @@ const completeQuest = async(req,res) => {
         console.log("database connected!");
 
         const questUpdated = await db.collection("quests").updateOne({_id: req.params.id}, {$set: {completed: true, completedAt: date}});
-        console.log(req.params.id, questUpdated);
 
-        // TO DO: Add quest task complete to user data
+        // Stop process if quest was not updated
+        if(!questUpdated){
+            client.close();
+            console.log("database disconnected!")
+            return res.status(400).json({status:400, data:questUpdated, message: "ERROR: Quest not updated."});
+        }
 
-        questUpdated
-        ? res.status(201).json({status:200, data:questUpdated, message: "SUCCESS: Quest deleted succesfully."})
-        : res.status(500).json({status:500, data:questUpdated, message: "ERROR: Internal server error."});   
+        // Get Owner details to award task points as well. 
+        // Get Participant details to award karma points and task points.
+        const questDetails = await db.collection("quests").findOne({_id: req.params.id});
+
+        const ownerUpdate = await db.collection("users").updateOne({_id: questDetails.ownerId}, {$inc: {taskPoints: questDetails.difficulty * 10}, $set: {updatedAt: date}});
+
+        // Stop process if owner was not updated
+        if(ownerUpdate === null){
+            client.close();
+            console.log("database disconnected!")
+            return res.status(400).json({status:400, data:ownerUpdate, message: "ERROR: Owner was not awarded task points"});
+        }
+
+        const updatePromise = [];
+        const karma = questDetails.karma/questDetails.participantIds.length;
+
+        questDetails.participantIds.forEach((participant) => {
+            updatePromise.push(db.collection("users").updateOne(
+                {_id: participant} , {$inc: {karma: karma}, $inc: {taskPoints: questDetails.difficulty * 10}, $set: {updatedAt: date}}));
+        })
+
+        Promise.all(updatePromise)
+        .then((response) => {
+            client.close();
+            console.log("database disconnected!")
+
+            console.log(response);
+            if (response.every(i => i !== null)){
+                res.status(200).json({status:200, data:response, message: "SUCCESS: Quest has been marked as complete."});
+            } else {
+                res.status(400).json({status:400, data:response, message: "ERROR: Users were not awarded task points and karma."});
+            }
+        })
     }catch(err){
         console.log(err);
-        res.status(500).json({status:500, message: `ERROR: Internal server error.`});
-    } finally {
         client.close();
         console.log("database disconnected!")
+        res.status(500).json({status:500, message: `ERROR: Internal server error.`});
     }
 }
